perf: Optimize chunk downloading using Copy() instead of Read() and Write()
io.Copy() doesn't create buffers for each chunk that need to be garbage collected later. Instead, it reuses the same buffer across chunks when writing. It also allows for streaming the file to disk instead of reading it all into memory at once, allowing for handling of larger chunk sizes.